refactor: 학생회 게시물 패키지 이름 수정
