+++ src/main/java/com/campus/campus/domain/councilpost/application/dto/response/NormalizedDateTime.java
@@ -1,0 +1,9 @@
+package com.campus.campus.domain.councilpost.application.dto.response;
+
+import java.time.LocalDateTime;
+
+public record NormalizedDateTime(
+	LocalDateTime startDateTime,
+	LocalDateTime endDateTime
+) {
+}
