+++ src/main/java/com/campus/campus/domain/councilpost/application/exception/EventStartDateTimeRequiredException.java
@@ -1,0 +1,9 @@
+package com.campus.campus.domain.councilpost.application.exception;
+
+import com.campus.campus.global.common.exception.ApplicationException;
+
+public class EventStartDateTimeRequiredException extends ApplicationException {
+	public EventStartDateTimeRequiredException() {
+		super(ErrorCode.EVENT_START_DATETIME_REQUIRED);
+	}
+}
