+++ src/main/java/com/campus/campus/domain/councilpost/application/exception/NotPostWriterException.java
@@ -1,0 +1,9 @@
+package com.campus.campus.domain.councilpost.application.exception;
+
+import com.campus.campus.global.common.exception.ApplicationException;
+
+public class NotPostWriterException extends ApplicationException {
+	public NotPostWriterException() {
+		super(ErrorCode.NOT_POST_WRITER);
+	}
+}
