+++ src/main/java/com/campus/campus/domain/councilpost/application/exception/PartnershipDateRequiredException.java
@@ -1,0 +1,9 @@
+package com.campus.campus.domain.councilpost.application.exception;
+
+import com.campus.campus.global.common.exception.ApplicationException;
+
+public class PartnershipDateRequiredException extends ApplicationException {
+	public PartnershipDateRequiredException() {
+		super(ErrorCode.PARTNERSHIP_DATE_REQUIRED);
+	}
+}
