+++ src/main/java/com/campus/campus/domain/councilpost/application/exception/PostNotFoundException.java
@@ -1,0 +1,9 @@
+package com.campus.campus.domain.councilpost.application.exception;
+
+import com.campus.campus.global.common.exception.ApplicationException;
+
+public class PostNotFoundException extends ApplicationException {
+	public PostNotFoundException() {
+		super(ErrorCode.POST_NOT_FOUND);
+	}
+}
