+++ src/main/java/com/campus/campus/domain/councilpost/application/exception/ThumbnailRequiredException.java
@@ -1,0 +1,9 @@
+package com.campus.campus.domain.councilpost.application.exception;
+
+import com.campus.campus.global.common.exception.ApplicationException;
+
+public class ThumbnailRequiredException extends ApplicationException {
+	public ThumbnailRequiredException() {
+		super(ErrorCode.THUMBNAIL_REQUIRED);
+	}
+}
