+++ src/main/java/com/campus/campus/domain/councilpost/domain/entity/ThumbnailIcon.java
@@ -1,0 +1,9 @@
+package com.campus.campus.domain.councilpost.domain.entity;
+
+public enum ThumbnailIcon {
+	CAFE,
+	EVENT,
+	NOTICE,
+	FOOD,
+	SPORTS
+}
