--- src/main/java/com/campus/campus/domain/studentcouncilpost/application/dto/response/NormalizedDateTime.java
@@ -1,9 +1,0 @@
-package com.campus.campus.domain.studentcouncilpost.application.dto.response;
-
-import java.time.LocalDateTime;
-
-public record NormalizedDateTime(
-	LocalDateTime startDateTime,
-	LocalDateTime endDateTime
-) {
-}
