--- src/main/java/com/campus/campus/domain/studentcouncilpost/application/exception/NotPostWriterException.java
@@ -1,9 +1,0 @@
-package com.campus.campus.domain.studentcouncilpost.application.exception;
-
-import com.campus.campus.global.common.exception.ApplicationException;
-
-public class NotPostWriterException extends ApplicationException {
-	public NotPostWriterException() {
-		super(ErrorCode.NOT_POST_WRITER);
-	}
-}
