--- src/main/java/com/campus/campus/domain/studentcouncilpost/application/exception/PostImageLimitExceededException.java
@@ -1,9 +1,0 @@
-package com.campus.campus.domain.studentcouncilpost.application.exception;
-
-import com.campus.campus.global.common.exception.ApplicationException;
-
-public class PostImageLimitExceededException extends ApplicationException {
-	public PostImageLimitExceededException() {
-		super(ErrorCode.POST_IMAGE_LIMIT_EXCEEDED);
-	}
-}
