--- src/main/java/com/campus/campus/domain/studentcouncilpost/application/exception/ThumbnailRequiredException.java
@@ -1,9 +1,0 @@
-package com.campus.campus.domain.studentcouncilpost.application.exception;
-
-import com.campus.campus.global.common.exception.ApplicationException;
-
-public class ThumbnailRequiredException extends ApplicationException {
-	public ThumbnailRequiredException() {
-		super(ErrorCode.THUMBNAIL_REQUIRED);
-	}
-}
