--- src/main/java/com/campus/campus/domain/studentcouncilpost/domain/entity/ThumbnailIcon.java
@@ -1,9 +1,0 @@
-package com.campus.campus.domain.studentcouncilpost.domain.entity;
-
-public enum ThumbnailIcon {
-	CAFE,
-	EVENT,
-	NOTICE,
-	FOOD,
-	SPORTS
-}
